Tidy: fix v2 linting errors and move Tiled parser bits into modules
