+++ v3/src/gameobjects/tilemap/parsers/parsetiledjson/BuildTilesetIndex.js
@@ -1,0 +1,52 @@
+// Master list of tiles -> x, y, index in tileset
+var BuildTilesetIndex = function (mapData)
+{
+    var tiles = [];
+
+    for (var i = 0; i < mapData.tilesets.length; i++)
+    {
+        var set = mapData.tilesets[i];
+
+        var x = set.tileMargin;
+        var y = set.tileMargin;
+
+        var count = 0;
+        var countX = 0;
+        var countY = 0;
+
+        for (var t = set.firstgid; t < set.firstgid + set.total; t++)
+        {
+            //  Can add extra properties here as needed
+            tiles[t] = [ x, y, i ];
+
+            x += set.tileWidth + set.tileSpacing;
+
+            count++;
+
+            if (count === set.total)
+            {
+                break;
+            }
+
+            countX++;
+
+            if (countX === set.columns)
+            {
+                x = set.tileMargin;
+                y += set.tileHeight + set.tileSpacing;
+
+                countX = 0;
+                countY++;
+
+                if (countY === set.rows)
+                {
+                    break;
+                }
+            }
+        }
+    }
+
+    return tiles;
+};
+
+module.exports = BuildTilesetIndex;
